feat: #215 Return Latest Entries and refactor Time Entries
